perf: 访问量统计功能优化

+++ blog-admin/src/main/java/cn/ahzoo/admin/model/vo/ViewStatisticsVO.java
@@ -11,6 +11,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class ViewStatisticsVO implements Serializable {
     private static final long serialVersionUID = 1L;
-    Integer pv;
-    Integer uv;
+    Long pv;
+    Long uv;
 }
